rtemslwip: Move ZynqMP servicing thread to core 0
There have been some SMP compatibility issues seen on ZynqMP systems,
even with the recent patch to manage the active pbuf pool. It was
confirmed with Xilinx that this driver was never intended to run with
multiple cores available even though lwIP has multi-core support and
there doesn't appear to be any mention of this limitation in public
documentation. This moves the ZynqMP hardware servicing thread to core 0
where the ISR server runs interrupt handlers.