embeddedsw/xemacps: Remove multicast listener limit
This removes the limit on the number of multicast addressses that can be
listened to simultaneously. This was erroneously limited to the number
of specific address registers available to be used for local MAC
addresses. Any number of multicast addresses can now be used, but there
is a known bug in the implementation that allows a multicast address
removal to inadvertedly disable other multicast listeners if there is a
hash collision. The correct way to manage these hash registers is to
maintain a list of multicast addresses and fully regenerate the hash
every time a multicast address is added or removed.

Closes #10